+++ frontend-v2/src/components/buttons/ExportButton.tsx
@@ -1,10 +1,12 @@
 import type { ButtonProps } from "@/types/props"
 
 import { Download } from "lucide-react"
+import { useTranslations } from "next-intl"
 
 import { Button } from "@/components/ui/button"
 
 const ExportButton = ({ handleClick }: ButtonProps) => {
+  const t = useTranslations("Actions");
   return (
     <Button
       onClick={handleClick}
@@ -13,7 +15,7 @@ const ExportButton = ({ handleClick }: ButtonProps) => {
       className="gap-2"
     >
       <Download className="h-4 w-4" />
-      Export
+      {t("export")}
     </Button>
   )
 }